Revisions for pipeline integration and separate annotations

+++ Makefile
@@ -21,7 +21,7 @@
 # along with this program; If not, see <http://www.gnu.org/licenses/>.                                     #
 ############################################################################################################
 
-SHELL = /usr/bin/bash
+SHELL = /usr/bin/env bash
 
 
 DESTDIR              ?=
@@ -58,7 +58,8 @@ DOCUMENTATIONS       := $(shell find License.txt changeLog.txt commandLineOption
 VC_FLAG              := $(DESTDIR)$(PYTHONDIR)/variantconvert/pipinstall.flag
 VC_VERSION           := 2.0.1
 VC_CONFIGDIR         := $(DESTDIR)$(PYTHONDIR)/variantconvert/src/variantconvert/configs
-
+USEANNODIR           := #flag whether separate annotation resources directory needed (e.g. for HPC environvment)
+EXRP_FILE            := #optional filepath for previously downloaded rest-prioritiser
 
 # make install
 .PHONY: install
@@ -82,8 +83,8 @@ install-display:
 
 install-documentationlight: $(DOCUMENTATIONS)
 	@echo ""
-	$(MV) $^ $(DESTDIR)$(DOCDIR)/$(ANNOTSV)
-	$(MV) $(TCLDIRDISTRIBUTED) $(TCLDIR)
+	$(CP) $^ $(DESTDIR)$(DOCDIR)/$(ANNOTSV)
+	$(CPDIR) $(TCLDIRDISTRIBUTED) $(TCLDIR)
 
 install-configfile: $(CONFIGFILE)
 	@echo ""
@@ -116,7 +117,13 @@ install-variantconvert:
 	@echo ""
 	@echo "variantconvert installation"
 	@echo "---------------------------"
+	
+	@if [ -d $(DESTDIR)$(PYTHONDIR)/variantconvert ]; then \
+		echo "variantconvert directory found; purging locally before re-installing."; \
+		rm -rf $(DESTDIR)$(PYTHONDIR)/variantconvert/; \
+	fi
 	git clone --branch $(VC_VERSION) https://github.com/SamuelNicaise/variantconvert.git $(DESTDIR)$(PYTHONDIR)/variantconvert/
+
 	touch $(VC_FLAG)
 	chmod 777 $(VC_FLAG)
 	pip3 install -e $(DESTDIR)$(PYTHONDIR)/variantconvert/. > ./tmp.variantconvert.txt 2>&1 \
@@ -179,7 +186,8 @@ install-done:
 # make install_organism_annotations
 install-all-annotations: install-human-annotation install-mouse-annotation                                     
 
-install-human-annotation: Annotations_Human_$(HUMANVERSION).tar.gz install-exomiser
+install-human-annotation: install-exomiser $(if $(USEANNODIR),,Annotations_Human_$(HUMANVERSION).tar.gz)
+ifndef USEANNODIR
 	@echo ""
 	@echo "Installation of human annotation:"
 	@echo ""
@@ -188,18 +196,36 @@ install-human-annotation: Annotations_Human_$(HUMANVERSION).tar.gz install-exomi
 	$(CHMOD) $(DESTDIR)$(SHAREDIR)/$(ANNOTSV)/Annotations_*
 	@echo ""
 	@echo "--> Human annotation installed"
+else
+	@echo ""
+	@echo "Flag for custom annotationDir; skipping local install of human annotations"
+	@echo ""
+endif
 
-install-exomiser-1: 2406_phenotype.zip
+install-exomiser-1: $(if $(USEANNODIR),,2406_phenotype.zip)
 	@echo ""
 	@echo "Installation of Exomiser data:"
 	@echo ""
+	
+ifndef USEANNODIR
 	$(MKDIR) -p $(DESTDIR)$(SHAREDIR)/$(ANNOTSV)/Annotations_Exomiser/2406
 	unzip 2406_phenotype.zip -d $(DESTDIR)$(SHAREDIR)/$(ANNOTSV)/Annotations_Exomiser/2406/
 	$(RM) -rf 2406_phenotype.zip
-	curl -C - -LO https://github.com/exomiser/Exomiser/releases/download/14.1.0/exomiser-rest-prioritiser-14.1.0.jar
+else
+	@echo ""
+	@echo "Flag for custom annotationDir; skipped Exomiser phenotypes local installation"
+	@echo ""
+endif
+	
 	$(MKDIR) -p $(DESTDIR)$(JARDIR)
+ifndef EXRP_FILE
+	curl -C - -LO https://github.com/exomiser/Exomiser/releases/download/14.1.0/exomiser-rest-prioritiser-14.1.0.jar
 	install -p -m 0755 exomiser-rest-prioritiser-14.1.0.jar $(DESTDIR)$(JARDIR)/
 	$(RM) exomiser-rest-prioritiser-14.1.0.jar
+else
+	@echo "Custom rest-priotiser path provided; creating symlink"
+	ln -sf $(EXRP_FILE) $(DESTDIR)$(JARDIR)/$(notdir $(EXRP_FILE))
+endif
 
 install-exomiser-2:
 	install -D -p -m 0755 $(PROPERTIES) $(DESTDIR)$(ETCDIR)/$(ANNOTSV)
@@ -208,15 +234,22 @@ install-exomiser-3:
 	@echo ""
 	@echo "--> Exomiser data installed"
 
-install-mouse-annotation: Annotations_Mouse_$(MOUSEVERSION).tar.gz 
+install-mouse-annotation: $(if $(USEANNODIR),,Annotations_Mouse_$(MOUSEVERSION).tar.gz) 
+ifndef USEANNODIR
 	@echo ""
 	@echo "Installation of mouse annotation:"
-	@echo ""
+	@echo ""	
 	$(MKDIR) $(DESTDIR)$(SHAREDIR)/$(ANNOTSV)/
 	tar -xf Annotations_Mouse_$(MOUSEVERSION).tar.gz -C $(DESTDIR)$(SHAREDIR)/$(ANNOTSV)/
 	$(RM) -rf Annotations_Mouse_$(MOUSEVERSION).tar.gz
 	@echo ""
 	@echo "--> Mouse annotation installed"
+else
+	@echo ""
+	@echo "Flag for custom annotationDir; skipping local install of mouse annotations"
+	@echo ""
+endif
+
 
 Annotations_%.tar.gz:
 	@echo ""
